requirements.txt
fastapi
uvicorn[standard]
openai
PyMuPDF
PyYAML
openpyxl
pydantic
